fix CLI run script, restructure acemcli package
