+++ bs/src/acemcli/cli.py
@@ -1,20 +1,22 @@
 from __future__ import annotations
-import sys, logging
+import sys
+import logging
 from pathlib import Path
 from urllib.parse import urlparse
-from acmecli.logging_setup import setup_logging
-from acmecli.config import load_config
-from acmecli.orchestrator import compute_all, to_ndjson
-from acmecli.models import Category
+
+from acemcli.logging_setup import setup_logging
+from acemcli.config import load_config
+from acemcli.orchestrator import compute_all, to_ndjson
+from acemcli.models import Category  # should be Literal["model","dataset","code"]
 
 log = logging.getLogger(__name__)
 
 def infer_category(url: str) -> Category:
     if "/datasets/" in url:
-        return "DATASET"
+        return "dataset"
     if url.startswith("https://huggingface.co/"):
-        return "MODEL"
-    return "CODE"
+        return "model"
+    return "code"
 
 def _is_valid_url(s: str) -> bool:
     try:
@@ -26,43 +28,41 @@ def _is_valid_url(s: str) -> bool:
 def main(url_file: str) -> int:
     setup_logging()
     cfg = load_config()
-    log.info("starting run: workers=%d log_level=%d", cfg.workers, cfg.log_level)
+    log.info("starting run: workers=%s log_level=%s", getattr(cfg, "workers", "?"), getattr(cfg, "log_level", "?"))
 
     p = Path(url_file)
-    if not p.is_absolute() or not p.exists():
-        print("URL_FILE must be an absolute path to an existing file", file=sys.stderr)
+    # Accept relative paths too; resolve to absolute
+    if not p.exists():
+        print("URL_FILE must point to an existing file", file=sys.stderr)
         return 1
+    p = p.resolve()
 
     # read, strip, drop comments and empties
     raw_lines = p.read_text(encoding="utf-8").splitlines()
-    cleaned = []
+    lines = []
+    seen = set()
     for ln in raw_lines:
         s = ln.strip()
         if not s or s.startswith("#"):
             continue
-        cleaned.append(s)
-
-    # dedupe while preserving order
-    seen = set()
-    lines: list[str] = []
-    for s in cleaned:
         if s not in seen:
             seen.add(s)
             lines.append(s)
 
-    # build (url, category) pairs, model-only for compute/output
+    # build (url, category) pairs; compute only for models per spec
     pairs: list[tuple[str, Category]] = []
-    for ln in lines:
-        if not _is_valid_url(ln):
-            log.warning("skipping invalid URL: %s", ln)
+    for s in lines:
+        if not _is_valid_url(s):
+            log.warning("skipping invalid URL: %s", s)
             continue
-        cat = infer_category(ln)
-        if cat == "MODEL":
-            pairs.append((ln, cat))
+        cat = infer_category(s)
+        if cat == "model":
+            pairs.append((s, cat))
 
     results, errors = compute_all(pairs)
     for res in results:
-        sys.stdout.buffer.write(to_ndjson(res))
+        # print one NDJSON line per model
+        print(to_ndjson(res))
 
     if errors:
         print(f"{len(errors)} URL(s) failed. See logs for details.", file=sys.stderr)
@@ -71,6 +71,6 @@ def main(url_file: str) -> int:
 
 if __name__ == "__main__":
     if len(sys.argv) < 2:
-        print("Usage: python -m acmecli.cli /absolute/path/to/URL_FILE", file=sys.stderr)
+        print("Usage: python -m acemcli.cli /path/to/URL_FILE", file=sys.stderr)
         raise SystemExit(1)
     raise SystemExit(main(sys.argv[1]))
