+++ bs/src/acemcli/logging_setup.py
@@ -1,17 +1,18 @@
 from __future__ import annotations
 import logging
-from acmecli.config import load_config
+from acemcli.config import load_config
 
 def setup_logging() -> None:
     cfg = load_config()
     level_map = {0: logging.CRITICAL, 1: logging.INFO, 2: logging.DEBUG}
-    lvl = level_map.get(cfg.log_level, logging.CRITICAL)
+    lvl = level_map.get(getattr(cfg, "log_level", 0), logging.CRITICAL)
 
     handlers = []
-    if cfg.log_file:
+    if getattr(cfg, "log_file", None):
         fh = logging.FileHandler(cfg.log_file, encoding="utf-8")
         fh.setLevel(lvl)
         handlers.append(fh)
+
     sh = logging.StreamHandler()
     sh.setLevel(lvl)
     handlers.append(sh)
@@ -20,5 +21,5 @@ def setup_logging() -> None:
         level=lvl,
         format="%(asctime)s [%(levelname)s] %(name)s: %(message)s",
         handlers=handlers,
-        force=True,  # ensure repeat init works in tests
+        force=True,
     )
