--- bs/src/metrics/__init__.py
@@ -1,6 +1,0 @@
-from . import hf_api
-from . import size_score
-from . import dataset_code_score
-from . import performance_claims
-from . import local_repo
-from . import dataset_quality
