+++ input/test_valid_urls.txt
@@ -1,0 +1,1 @@
+https://huggingface.co/google-bert/bert-base-uncased
