--- src/init.py
@@ -1,13 +1,0 @@
-# src/metrics/__init__.py
-__all__ = [
-    "busFactor",
-    "hf_api",
-    "dataset_quality",
-    "size_score",
-    "dataset_code_score",
-    "licenseCheck",
-    "local_repo",
-    "performance_claims",
-    "codeCheck",
-]
-# No eager imports here. Tests will import submodules directly.
